+++ projects/SmallTheftAuto/Assets/Main/Scripts/UI/UIHealthbarScript_ML.cs
@@ -176,7 +176,7 @@ public class UIHealthbarScript_ML : MonoBehaviour
         }
     }
     
-    private void DecrementHeart(float damageAmount)
+    private void DecrementHeart(int damageAmount)
     {
         for (int i = 9; i >= 0; i--)
         {
